Integrate runtime (which was mostly just dispatching to other delegates) into DefaultContext
