+++ pi4j-core/src/main/java/com/pi4j/context/impl/MutableRegistry.java
@@ -1,4 +1,4 @@
-package com.pi4j.runtime;
+package com.pi4j.context.impl;
 
 /*-
  * #%L
